@@ -6,10 +6,10 @@ authors = [{ name = "Visesh Prasad", email = "viseshrprasad@gmail.com" }]
 readme = "README.md"
 license = { text = "MIT" }
 keywords = ["workedon", "work", "worklog", "log", "journal"]
-requires-python = ">=3.9,<4.0"
+requires-python = ">=3.10,<3.15"
 dependencies = [
     "click>=8.1.1",
-    "dateparser>=1.1.4,<1.2.1",  # behavior changes in 1.2.1
+    "dateparser>=1.2.2, <1.3.0",  # note: previous behavior changes in 1.2.1
     "tzlocal>=4.0.0",
     "click-default-group>=1.2.4",
     "peewee>=3.18.1",
@@ -25,11 +25,11 @@ classifiers = [
     "Topic :: Software Development :: Libraries :: Python Modules",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: 3.14",
 ]
 
 [project.urls]
@@ -147,7 +147,7 @@ source = ["workedon"]
 show_contexts = true
 
 [tool.ruff]
-target-version = "py39"
+target-version = "py310"
 line-length = 100
 fix = true
 exclude = [".venv", "__init__.py"]
@@ -209,7 +209,7 @@ force-sort-within-sections = true
 
 [tool.black]
 line-length = 100
-target-version = ["py39"]
+target-version = ["py310"]
 
 [tool.codespell]
 ignore-words = "codespell.txt"
@@ -230,4 +230,4 @@ exclude = ["venv/", ".venv/"]
 exclude = ["venv/.*", ".venv/.*", "tests/.*"]
 
 [tool.pre-commit.default_language_versions]
-python = "3.13"
+python = "3.14"
